Rename transforms subdir to transform
