+++ src/ast/transform/AstTransformer.cpp
@@ -14,7 +14,7 @@
  *
  ***********************************************************************/
 
-#include "ast//transforms/AstTransformer.h"
+#include "ast//transform/AstTransformer.h"
 #include "ErrorReport.h"
 #include "ast/AstTranslationUnit.h"
 #include <chrono>
