+++ src/ast/transform/ComponentInstantiationTransformer.h
@@ -14,7 +14,7 @@
 
 #pragma once
 
-#include "ast/transforms/AstTransformer.h"
+#include "ast/transform/AstTransformer.h"
 #include <string>
 
 namespace souffle {
