+++ src/ast/transform/InlineRelationsTransformer.cpp
@@ -29,7 +29,7 @@
 #include "ast/AstTranslationUnit.h"
 #include "ast/AstUtils.h"
 #include "ast/AstVisitor.h"
-#include "ast/transforms/AstTransforms.h"
+#include "ast/transform/AstTransforms.h"
 #include "utility/MiscUtil.h"
 #include <algorithm>
 #include <cassert>
