+++ src/ast/transform/MagicSet.h
