+++ src/ast/transform/MinimiseProgramTransformer.cpp
@@ -25,7 +25,7 @@
 #include "ast/AstUtils.h"
 #include "ast/AstVisitor.h"
 #include "ast/analysis/AstIOTypeAnalysis.h"
-#include "ast/transforms/AstTransforms.h"
+#include "ast/transform/AstTransforms.h"
 #include "utility/ContainerUtil.h"
 #include <algorithm>
 #include <cassert>
