+++ src/main.cpp
@@ -34,11 +34,11 @@
 #include "ast/AstTranslationUnit.h"
 #include "ast/analysis/AstTypeAnalysis.h"
 #include "ast/analysis/PrecedenceGraph.h"
-#include "ast/transforms/AstComponentChecker.h"
-#include "ast/transforms/AstPragmaChecker.h"
-#include "ast/transforms/AstSemanticChecker.h"
-#include "ast/transforms/AstTransforms.h"
-#include "ast/transforms/ComponentInstantiationTransformer.h"
+#include "ast/transform/AstComponentChecker.h"
+#include "ast/transform/AstPragmaChecker.h"
+#include "ast/transform/AstSemanticChecker.h"
+#include "ast/transform/AstTransforms.h"
+#include "ast/transform/ComponentInstantiationTransformer.h"
 #include "config.h"
 #include "profile/Tui.h"
 #include "utility/FileUtil.h"
