+++ src/tests/ast_utils_test.cpp
@@ -30,7 +30,7 @@
 #include "ast/AstTranslationUnit.h"
 #include "ast/AstUtils.h"
 #include "ast/analysis/AstGroundAnalysis.h"
-#include "ast/transforms/AstTransforms.h"
+#include "ast/transform/AstTransforms.h"
 #include "utility/StringUtil.h"
 #include <algorithm>
 #include <map>
